Fix DiscoveryCollectionsStorageCrs look up of global crs
DiscoveryCollectionsStorageCrs asserts that a `/collections` document
complies with

```
Abstract Test 3: /conf/crs/storageCrs
Test Purpose: Verify that the storage CRS identifier is a valid value

Test Method: For each collection object that includes a storageCrs property in the paths /collections and
/collections/{collectionId}, validate that the string is also found in the crs property of the collection or, in case
the crs property includes a value #/crs, in the global list of CRSs.
```

but doesn't expand the collection's `#/crs` JSON pointer to the collections
document crs list.

This patch fixes it and adds unit tests to cover all possible cases.